// types.ts
export type Position = {
  x: number;
  y: number;
};

export type Difficulty = 'Easy' | 'Medium' | 'Hard';

export type ThemeColors = {
  primary: string;
  primaryHover: string;
  primaryRing: string;
  shadow: string;
  snakeHead: string;
  snakeBody: string;
  food: string;
  background: string;
  grid: string;
};